* Finnally implemented new sorting system What I was doing;
Trying to take over the sort completely to arbitarily move stuff to the exact position I wanted

What I should have been doing;

Simply set the "distance" of a object to the same as the distances of the object I want it next too, then let the existing sort do the work

This new system works fine, but has yet to be fully applied to all the scene objects in a nice way.
Once this "z index" system is done, it should tidy up a lot of the interface glitchs.

TODO:
Sort can be optimised a bit - currently we are calculating a real distance to camera even when using a "fake" one.